+++ itsm_frontend/src/modules/assets/api/index.ts
@@ -1,0 +1,2 @@
+// index.ts for assets api
+export {};
