+++ itsm_frontend/src/modules/assets/components/index.ts
@@ -1,0 +1,6 @@
+// itsm_frontend/src/modules/assets/components/index.ts
+export * from './AssetForm';
+export * from './AssetList';
+export * from './CategoryManagement';
+export * from './LocationManagement';
+export * from './VendorManagement';
