+++ itsm_frontend/src/modules/assets/hooks/index.ts
@@ -1,0 +1,2 @@
+// index.ts for assets hooks
+export {};
