+++ itsm_frontend/src/modules/assets/index.ts
@@ -1,7 +1,6 @@
-// itsm_frontend/src/modules/service-requests/index.ts
+// itsm_frontend/src/modules/assets/index.ts
 export * from './api';
 export * from './components';
-export * from './context';
 export * from './hooks';
 export * from './pages';
 export * from './types';
