+++ itsm_frontend/src/modules/assets/pages/index.ts
@@ -1,0 +1,2 @@
+// index.ts for assets pages
+export * from './AssetsPage';
