+++ itsm_frontend/src/modules/assets/types/index.ts
@@ -1,0 +1,2 @@
+// index.ts for assets types
+export * from './assetTypes';
