+++ itsm_frontend/src/modules/procurement/api/index.ts
@@ -1,0 +1,7 @@
+// itsm_frontend/src/modules/procurement/api/index.ts
+// This file can re-export functions from procurementApi.ts if it's moved here,
+// or serve as a barrel file if api logic is further broken down.
+// For now, it's a placeholder.
+// If procurementApi.ts is not moved into this folder, this might not be strictly necessary
+// depending on how API functions are accessed.
+export {};
