+++ itsm_frontend/src/modules/procurement/components/index.ts
@@ -1,0 +1,4 @@
+// itsm_frontend/src/modules/procurement/components/index.ts
+export * from './check-requests';
+export * from './purchase-orders';
+export * from './purchase-request-memos';
