+++ itsm_frontend/src/modules/procurement/components/purchase-orders/PurchaseOrderList.tsx
@@ -25,7 +25,7 @@ import EditIcon from '@mui/icons-material/Edit'; // Added EditIcon
 import { useAuth } from '../../../context/auth/useAuth';
 // import { useUI } from '../../../context/UIContext/useUI'; // Not strictly needed if only snackbar for error/success
 import { getPurchaseOrders } from '../../../api/procurementApi';
-import type { PurchaseOrder, GetPurchaseOrdersParams, PurchaseOrderStatus } from '../../../api/procurementApi';
+import type { PurchaseOrder, GetPurchaseOrdersParams, PurchaseOrderStatus } from '../../types';
 import { useNavigate } from 'react-router-dom';
 
 type Order = 'asc' | 'desc';
