+++ itsm_frontend/src/modules/procurement/components/purchase-orders/index.ts
@@ -1,0 +1,4 @@
+// itsm_frontend/src/modules/procurement/components/purchase-orders/index.ts
+export * from './PurchaseOrderDetailView';
+export * from './PurchaseOrderForm';
+export * from './PurchaseOrderList';
