+++ itsm_frontend/src/modules/procurement/components/purchase-request-memos/index.ts
@@ -1,0 +1,4 @@
+// itsm_frontend/src/modules/procurement/components/purchase-request-memos/index.ts
+export * from './PurchaseRequestMemoDetailView';
+export * from './PurchaseRequestMemoForm';
+export * from './PurchaseRequestMemoList';
