+++ itsm_frontend/src/modules/procurement/hooks/index.ts
@@ -1,0 +1,3 @@
+// itsm_frontend/src/modules/procurement/hooks/index.ts
+// Export any hooks here in the future
+export {};
