+++ itsm_frontend/src/modules/procurement/index.ts
@@ -1,0 +1,6 @@
+// itsm_frontend/src/modules/procurement/index.ts
+export * from './api'; // Assuming api functions might be exported from an index file here
+export * from './components';
+export * from './hooks';
+export * from './pages';
+export * from './types';
