+++ itsm_frontend/src/modules/procurement/pages/index.ts
@@ -1,0 +1,3 @@
+// itsm_frontend/src/modules/procurement/pages/index.ts
+export * from './ProcurementDashboardPage';
+// Add other page exports here if any are moved or created
