+++ itsm_frontend/src/modules/procurement/types/index.ts
@@ -1,0 +1,2 @@
+// itsm_frontend/src/modules/procurement/types/index.ts
+export * from './procurementTypes';
