+++ itsm_frontend/src/modules/service-requests/api/index.ts
@@ -1,0 +1,3 @@
+// itsm_frontend/src/modules/service-requests/api/index.ts
+// This file can export API-related functions if refactored here in the future.
+export {};
