+++ itsm_frontend/src/modules/service-requests/components/index.ts
@@ -1,0 +1,4 @@
+// itsm_frontend/src/modules/service-requests/components/index.ts
+export * from './ServiceRequestForm';
+export * from './ServiceRequestList';
+export * from './ServiceRequestPrintView';
