+++ itsm_frontend/src/modules/service-requests/hooks/index.ts
@@ -1,0 +1,2 @@
+// itsm_frontend/src/modules/service-requests/hooks/index.ts
+export * from './useServiceRequests';
