+++ itsm_frontend/src/modules/service-requests/pages/index.ts
@@ -1,0 +1,3 @@
+// itsm_frontend/src/modules/service-requests/pages/index.ts
+export * from './NewServiceRequestPage';
+export * from './ServiceRequestsPage';
