+++ itsm_frontend/src/modules/service-requests/types/index.ts
@@ -1,0 +1,2 @@
+// itsm_frontend/src/modules/service-requests/types/index.ts
+export * from './ServiceRequestTypes';
